+++ experimental/instrumentation-xhr/src/index.ts
@@ -1,0 +1,3 @@
+export { XHRInstrumentation } from './instrumentation';
+
+export { parseXHREvent } from './utils';